refactor: k8s wds deployment section
